pyproject.toml
[tool.poetry]
name = "2024-aws"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
amazon-braket-sdk = "^1.68.2"
qiskit = "^0.46.0"
qiskit-ionq = "^0.4.7"
matplotlib = "^3.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
